Revert "Embed Slack Inviter on Desktop"
This reverts commit 45402a006ab72683513f0c481174be140ba4bb1e.

Turns out even with a LetsEncrypt certificate heroku requires you to
purchase their hobby dyno and I'm too cheap for that right now.

That's an issue because we'd be embedding HTTP on a HTTPS connection,
something browsers don't allow. I only noticed this when it was live
because I'm testing on HTTP locally.